editor/libeditor/tests/test_bug432225.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=432225
-->
<head>
  <title>Test for Bug 432225</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>  
  <script src="spellcheck.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=432225">Mozilla Bug 432225</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 432225 **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

var gMisspeltWords = [];

function getEdit() {
  return document.getElementById('edit');
}

function editDoc() {
  return getEdit().contentDocument;
}

function getEditor() {
  var Ci = SpecialPowers.Ci;
  var win = editDoc().defaultView;
  var editingSession = SpecialPowers.wrap(win)
                                    .QueryInterface(Ci.nsIInterfaceRequestor)
                                    .getInterface(Ci.nsIWebNavigation)
                                    .QueryInterface(Ci.nsIInterfaceRequestor)
                                    .getInterface(Ci.nsIEditingSession);
  return editingSession.getEditorForWindow(win);
}

function runTest() {
  editDoc().designMode = "on";
  setTimeout(function() { addWords(100); }, 0);
}  
 
function addWords(aLimit) {
  if (aLimit == 0) {
    ok(isSpellingCheckOk(getEditor(), gMisspeltWords),
       "All misspellings accounted for.");
    SimpleTest.finish();
    return;
  }
  getEdit().focus();
  sendString('aa OK ');
  gMisspeltWords.push("aa");
  setTimeout(function() { addWords(aLimit-1); }, 0);
}
</script>
</pre>

<iframe id="edit" width="200" height="100" src="about:blank"></iframe>

</body>
</html>
